backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# CORS middleware (included with FastAPI but listing for clarity)
# fastapi already includes starlette which has CORS support

# Google Gemini AI
google-generativeai==0.3.2

# Environment variables
python-dotenv==1.0.0

# PDF processing
PyMuPDF==1.23.8

# DOCX processing
python-docx==1.1.0

# HTTP client (if needed for API calls)
httpx==0.25.2

# Data validation and parsing (included with FastAPI via pydantic)
# pydantic is already included with FastAPI

# File handling and utilities
pathlib2==2.3.7

# Type hints support for older Python versions (if needed)
typing-extensions==4.8.0
